feat(server): migrate copilot provider sdk (#11584)
fix AI-15
fix AI-16
