+++ packages/backend/server/package.json
@@ -25,7 +25,9 @@
     "postinstall": "prisma generate"
   },
   "dependencies": {
-    "@ai-sdk/google": "^1.1.19",
+    "@ai-sdk/google": "^1.2.10",
+    "@ai-sdk/openai": "^1.3.9",
+    "@ai-sdk/perplexity": "^1.1.6",
     "@apollo/server": "^4.11.3",
     "@aws-sdk/client-s3": "^3.779.0",
     "@aws-sdk/s3-request-presigner": "^3.779.0",
@@ -73,7 +75,6 @@
     "date-fns": "^4.0.0",
     "dotenv": "^16.4.7",
     "eventemitter2": "^6.4.9",
-    "eventsource-parser": "^3.0.0",
     "express": "^5.0.1",
     "fast-xml-parser": "^5.0.0",
     "get-stream": "^9.0.1",
@@ -95,7 +96,6 @@
     "nestjs-cls": "^5.0.0",
     "nodemailer": "^6.9.16",
     "on-headers": "^1.0.2",
-    "openai": "^4.83.0",
     "piscina": "^5.0.0-alpha.0",
     "prisma": "^5.22.0",
     "react": "19.1.0",
