+++ packages/backend/server/src/plugins/copilot/context/embedding.ts
@@ -1,27 +1,39 @@
-import OpenAI from 'openai';
+import {
+  createOpenAI,
+  type OpenAIProvider as VercelOpenAIProvider,
+} from '@ai-sdk/openai';
+import { embedMany } from 'ai';
 
 import { Embedding } from '../../../models';
+import { OpenAIConfig } from '../providers/openai';
 import { EmbeddingClient } from './types';
 
 export class OpenAIEmbeddingClient extends EmbeddingClient {
-  constructor(private readonly client: OpenAI) {
+  readonly #instance: VercelOpenAIProvider;
+
+  constructor(config: OpenAIConfig) {
     super();
+    this.#instance = createOpenAI({
+      apiKey: config.apiKey,
+      baseURL: config.baseUrl,
+    });
   }
 
-  async getEmbeddings(
-    input: string[],
-    signal?: AbortSignal
-  ): Promise<Embedding[]> {
-    const resp = await this.client.embeddings.create(
-      {
-        input,
-        model: 'text-embedding-3-large',
-        dimensions: 1024,
-        encoding_format: 'float',
-      },
-      { signal }
-    );
-    return resp.data.map(e => ({ ...e, content: input[e.index] }));
+  async getEmbeddings(input: string[]): Promise<Embedding[]> {
+    const modelInstance = this.#instance.embedding('text-embedding-3-large', {
+      dimensions: 1024,
+    });
+
+    const { embeddings } = await embedMany({
+      model: modelInstance,
+      values: input,
+    });
+
+    return Array.from(embeddings.entries()).map(([index, embedding]) => ({
+      index,
+      embedding,
+      content: input[index],
+    }));
   }
 }
