+++ packages/backend/server/src/plugins/copilot/providers/gemini.ts
@@ -4,14 +4,10 @@ import {
 } from '@ai-sdk/google';
 import {
   AISDKError,
-  type CoreAssistantMessage,
-  type CoreUserMessage,
-  FilePart,
   generateObject,
   generateText,
   JSONParseError,
   streamText,
-  TextPart,
 } from 'ai';
 
 import {
@@ -29,34 +25,14 @@ import {
   CopilotTextToTextProvider,
   PromptMessage,
 } from './types';
+import { chatToGPTMessage } from './utils';
 
 export const DEFAULT_DIMENSIONS = 256;
-
-const SIMPLE_IMAGE_URL_REGEX = /^(https?:\/\/|data:image\/)/;
-const FORMAT_INFER_MAP: Record<string, string> = {
-  pdf: 'application/pdf',
-  mp3: 'audio/mpeg',
-  wav: 'audio/wav',
-  png: 'image/png',
-  jpeg: 'image/jpeg',
-  jpg: 'image/jpeg',
-  webp: 'image/webp',
-  txt: 'text/plain',
-  md: 'text/plain',
-  mov: 'video/mov',
-  mpeg: 'video/mpeg',
-  mp4: 'video/mp4',
-  avi: 'video/avi',
-  wmv: 'video/wmv',
-  flv: 'video/flv',
-};
 
 export type GeminiConfig = {
   apiKey: string;
   baseUrl?: string;
 };
-
-type ChatMessage = CoreUserMessage | CoreAssistantMessage;
 
 export class GeminiProvider
   extends CopilotProvider<GeminiConfig>
@@ -84,67 +60,6 @@ export class GeminiProvider
       apiKey: this.config.apiKey,
       baseURL: this.config.baseUrl,
     });
-  }
-
-  private inferMimeType(url: string) {
-    if (url.startsWith('data:')) {
-      return url.split(';')[0].split(':')[1];
-    }
-    const extension = url.split('.').pop();
-    if (extension) {
-      return FORMAT_INFER_MAP[extension];
-    }
-    return undefined;
-  }
-
-  protected async chatToGPTMessage(
-    messages: PromptMessage[]
-  ): Promise<[string | undefined, ChatMessage[], any]> {
-    const system =
-      messages[0]?.role === 'system' ? messages.shift() : undefined;
-    const schema = system?.params?.schema;
-
-    // filter redundant fields
-    const msgs: ChatMessage[] = [];
-    for (let { role, content, attachments, params } of messages.filter(
-      m => m.role !== 'system'
-    )) {
-      content = content.trim();
-      role = role as 'user' | 'assistant';
-      const mimetype = params?.mimetype;
-      if (Array.isArray(attachments)) {
-        const contents: (TextPart | FilePart)[] = [];
-        if (content.length) {
-          contents.push({
-            type: 'text',
-            text: content,
-          });
-        }
-
-        for (const url of attachments) {
-          if (SIMPLE_IMAGE_URL_REGEX.test(url)) {
-            const mimeType =
-              typeof mimetype === 'string' ? mimetype : this.inferMimeType(url);
-            if (mimeType) {
-              const data = url.startsWith('data:')
-                ? await fetch(url).then(r => r.arrayBuffer())
-                : new URL(url);
-              contents.push({
-                type: 'file' as const,
-                data,
-                mimeType,
-              });
-            }
-          }
-        }
-
-        msgs.push({ role, content: contents } as ChatMessage);
-      } else {
-        msgs.push({ role, content });
-      }
-    }
-
-    return [system?.content, msgs, schema];
   }
 
   protected async checkParams({
@@ -223,7 +138,7 @@ export class GeminiProvider
     try {
       metrics.ai.counter('chat_text_calls').add(1, { model });
 
-      const [system, msgs, schema] = await this.chatToGPTMessage(messages);
+      const [system, msgs, schema] = await chatToGPTMessage(messages);
 
       const modelInstance = this.#instance(model, {
         structuredOutputs: Boolean(options.jsonMode),
@@ -274,7 +189,7 @@ export class GeminiProvider
 
     try {
       metrics.ai.counter('chat_text_stream_calls').add(1, { model });
-      const [system, msgs] = await this.chatToGPTMessage(messages);
+      const [system, msgs] = await chatToGPTMessage(messages);
 
       const { textStream } = streamText({
         model: this.#instance(model),
