+++ packages/frontend/media-capture-playground/package.json
@@ -23,11 +23,9 @@
     "fs-extra": "^11.3.0",
     "lodash-es": "^4.17.21",
     "multer": "^1.4.5-lts.1",
-    "openai": "^4.85.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-markdown": "^10.0.0",
-    "rxjs": "^7.8.1",
     "socket.io": "^4.7.4",
     "socket.io-client": "^4.7.4",
     "swr": "^2.3.2",
